tmperature converted
The user will input a temperature in either Fahrenheit or Celsius and press a "convert" button. The converted temperature will then be displayed with the correct unit.
Your UI should include four main elements:
Input field for the user to type in their temperature - don't forget to validate that the input is a number!
Dropdown menu or radio button for the user to choose if they are inputting in Celsius or Fahrenheit
"Convert" button to trigger the temperature conversion
Display area for the final converted temperature and unit
For an extra challenge: You can add in a third conversion option to convert from and to Kelvin as well.